Move cloexec.* files out of common
These files are only used by the main program.
